Simplify GameLoop usage: rename GameLoop to Game
Now examples only need to define the main methods:

- event
- update
- render

Removed start/stop triggers for now.
